# HydraMind_v1/docs/API/openapi.yaml
openapi: 3.0.3
info:
  title: HydraMind v1 API
  description: |
    REST API for HydraMind v1 - Universal Cognitive Kernel for Intelligent Systems

    This API provides monitoring, control, and integration capabilities for HydraMind deployments.

    ## Authentication

    Currently, the API does not require authentication for development and testing.
    For production deployments, implement appropriate authentication mechanisms.

    ## Rate Limiting

    API endpoints are rate-limited to prevent abuse:
    - 100 requests per minute for health endpoints
    - 1000 requests per minute for metrics endpoints
    - 10 requests per minute for configuration changes

    ## Error Handling

    All endpoints return appropriate HTTP status codes:
    - 200: Success
    - 400: Bad Request (invalid parameters)
    - 404: Not Found (endpoint or resource not found)
    - 429: Too Many Requests (rate limit exceeded)
    - 500: Internal Server Error (server-side issues)
    - 503: Service Unavailable (system unhealthy)

  version: 1.0.0
  contact:
    name: HydraMind Support
    email: support@hydramind.dev
    url: https://hydramind.dev/support
  license:
    name: Proprietary
    url: https://hydramind.dev/license

servers:
  - url: http://localhost:8765
    description: Local development server
  - url: https://api.hydramind.dev/v1
    description: Production API server

components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
          description: Overall system health status
        modules:
          type: integer
          description: Number of active modules
          minimum: 0
        uptime:
          type: string
          description: System uptime in human-readable format
          example: "2h 15m 30s"
        health_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Overall health score (0.0 = completely unhealthy, 1.0 = perfectly healthy)
        details:
          type: object
          description: Detailed health information for each module
          additionalProperties:
            type: object
            properties:
              state:
                type: string
                enum: [uninitialized, initializing, running, stopping, stopped, error]
              uptime:
                type: number
                format: float
              message_count:
                type: integer
              error_count:
                type: integer
              memory_usage:
                type: integer
              cpu_usage:
                type: number
                format: float
              health_score:
                type: number
                format: float
      example:
        status: "healthy"
        modules: 8
        uptime: "2h 15m"
        health_score: 0.95
        details:
          temperature_monitor:
            state: "running"
            uptime: 7200.0
            message_count: 1500
            error_count: 0
            memory_usage: 52428800
            cpu_usage: 2.5
            health_score: 0.98

    MetricsResponse:
      type: object
      properties:
        timestamp:
          type: number
          format: float
          description: Unix timestamp when metrics were collected
        system:
          type: object
          properties:
            uptime:
              type: number
              format: float
            cpu_usage:
              type: number
              format: float
            memory_usage:
              type: integer
            disk_usage:
              type: number
              format: float
            network_io:
              type: object
              properties:
                bytes_sent:
                  type: integer
                bytes_received:
                  type: integer
        eventbus:
          type: object
          properties:
            messages_published:
              type: integer
            messages_dispatched:
              type: integer
            dispatch_errors:
              type: integer
            queue_size:
              type: integer
            subscribers:
              type: integer
            patterns:
              type: integer
        modules:
          type: object
          description: Per-module metrics
          additionalProperties:
            type: object
            properties:
              messages_processed:
                type: integer
              processing_time_avg:
                type: number
                format: float
              errors:
                type: integer
              memory_usage:
                type: integer
      example:
        timestamp: 1640995200.0
        system:
          uptime: 7200.0
          cpu_usage: 45.2
          memory_usage: 536870912
          disk_usage: 23.1
          network_io:
            bytes_sent: 1048576
            bytes_received: 2097152
        eventbus:
          messages_published: 15420
          messages_dispatched: 15380
          dispatch_errors: 2
          queue_size: 15
          subscribers: 12
          patterns: 8
        modules:
          temperature_monitor:
            messages_processed: 1500
            processing_time_avg: 0.023
            errors: 0
            memory_usage: 52428800

    EventQueryRequest:
      type: object
      properties:
        topic_pattern:
          type: string
          description: Topic pattern to match (supports wildcards like sensor/*)
          example: "sensor/*"
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 100
          description: Maximum number of events to return
        since:
          type: number
          format: float
          description: Return events since this Unix timestamp
        until:
          type: number
          format: float
          description: Return events until this Unix timestamp
        order:
          type: string
          enum: [asc, desc]
          default: desc
          description: Sort order for results

    EventQueryResponse:
      type: object
      properties:
        events:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              topic:
                type: string
              data:
                type: object
              timestamp:
                type: number
                format: float
              qos:
                type: integer
              key:
                type: string
                nullable: true
        total:
          type: integer
          description: Total number of events matching query
        returned:
          type: integer
          description: Number of events returned in this response

    EventPublishRequest:
      type: object
      required:
        - topic
        - data
      properties:
        topic:
          type: string
          description: Event topic to publish to
          example: "sensor/temperature"
        data:
          type: object
          description: Event payload (must be JSON-serializable)
          example: {"value": 23.5, "unit": "C"}
        qos:
          type: integer
          enum: [0, 1]
          default: 0
          description: Quality of service level
        key:
          type: string
          description: Optional key for deduplication/grouping

    EventPublishResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the event was published successfully
        message_id:
          type: string
          description: Unique identifier for the published event
        timestamp:
          type: number
          format: float
          description: When the event was published

    ModuleListResponse:
      type: array
      items:
        type: object
        properties:
          name:
            type: string
            description: Module name
          state:
            type: string
            enum: [uninitialized, initializing, running, stopping, stopped, error]
          uptime:
            type: number
            format: float
          message_count:
            type: integer
          error_count:
            type: integer
          subscriptions:
            type: array
            items:
              type: string
          health_score:
            type: number
            format: float

    ConfigurationResponse:
      type: object
      properties:
        version:
          type: string
          description: Configuration schema version
        server:
          type: object
          properties:
            enabled:
              type: boolean
            host:
              type: string
            port:
              type: integer
        logging:
          type: object
          properties:
            level:
              type: string
            json:
              type: boolean
            file_path:
              type: string
        features:
          type: object
          description: Enabled feature flags
        custom:
          type: object
          description: Custom configuration (sensitive values redacted)

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code for programmatic handling
        details:
          type: object
          description: Additional error context
        timestamp:
          type: number
          format: float

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (when enabled)

paths:
  /health:
    get:
      summary: Get system health status
      description: |
        Returns comprehensive health information for the HydraMind system including:
        - Overall system status
        - Individual module health
        - Performance metrics
        - Resource utilization
      operationId: getHealth
      tags:
        - Health
      responses:
        '200':
          description: System is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: System is unhealthy or unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /metrics:
    get:
      summary: Get system performance metrics
      description: |
        Returns detailed performance metrics including:
        - Event processing statistics
        - Resource utilization
        - Module performance data
        - System throughput measurements
      operationId: getMetrics
      tags:
        - Metrics
      responses:
        '200':
          description: Metrics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricsResponse'
        '500':
          description: Failed to retrieve metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /modules:
    get:
      summary: List all registered modules
      description: |
        Returns information about all currently registered modules including:
        - Module names and states
        - Performance statistics
        - Subscription patterns
        - Health information
      operationId: listModules
      tags:
        - Modules
      responses:
        '200':
          description: Modules retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleListResponse'
        '500':
          description: Failed to retrieve module information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /config:
    get:
      summary: Get current configuration
      description: |
        Returns the current system configuration with sensitive values redacted.
        Useful for debugging and monitoring configuration state.
      operationId: getConfiguration
      tags:
        - Configuration
      responses:
        '200':
          description: Configuration retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationResponse'
        '500':
          description: Failed to retrieve configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /bus/publish:
    post:
      summary: Publish an event to the event bus
      description: |
        Inject an event into the HydraMind event bus for testing, debugging,
        or integration purposes. The event will be processed by all subscribed modules.
      operationId: publishEvent
      tags:
        - Events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventPublishRequest'
      responses:
        '200':
          description: Event published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPublishResponse'
        '400':
          description: Invalid event data or topic
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /events/query:
    post:
      summary: Query historical events
      description: |
        Query the event store for historical events matching specified criteria.
        Supports topic pattern matching, time range filtering, and result limiting.
      operationId: queryEvents
      tags:
        - Events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventQueryRequest'
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventQueryResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No events found matching query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /shutdown:
    post:
      summary: Initiate graceful shutdown
      description: |
        Initiates a graceful shutdown of the HydraMind system. All modules will
        be stopped cleanly, resources will be released, and the system will exit.
      operationId: shutdownSystem
      tags:
        - System
      responses:
        '200':
          description: Shutdown initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "shutdown_initiated"
                  message:
                    type: string
                    example: "System will shut down gracefully"
        '409':
          description: Shutdown already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /:
    get:
      summary: Root endpoint
      description: |
        Returns basic information about the HydraMind API and system status.
        Useful for API discovery and basic connectivity testing.
      operationId: getRoot
      tags:
        - System
      responses:
        '200':
          description: API information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: "HydraMind Control Plane"
                  version:
                    type: string
                    example: "1.0.0"
                  status:
                    type: string
                    example: "operational"
                  documentation:
                    type: string
                    example: "https://docs.hydramind.dev"

tags:
  - name: Health
    description: System health monitoring and diagnostics
  - name: Metrics
    description: Performance metrics and system statistics
  - name: Modules
    description: Module management and monitoring
  - name: Configuration
    description: System configuration access and management
  - name: Events
    description: Event injection and querying capabilities
  - name: System
    description: System control and lifecycle management

externalDocs:
  description: Complete HydraMind Documentation
  url: https://docs.hydramind.dev